rename configs to constants
